feat(LineController): check line name length
