+++ docs/README.md
@@ -16,7 +16,7 @@
     - [ ] 등록된 역을 조회한다.
     
 - [x] 노선 이름을 입력받는다.
-    - [ ] 노선 이름이 2글자 이상인지 확인한다.
+    - [x] 노선 이름이 2글자 이상인지 확인한다.
     - [ ] 노선의 이름이 '선'으로 끝나는지 확인한다? (보류)
     - [ ] 노선이 이미 존재하는지 확인한다.
     - [ ] 노선의 상행 종점역과 하행 종점역을 입력받는다.
