--- src/main/java/subway/domain/Line.java
@@ -1,15 +1,0 @@
-package subway.domain;
-
-public class Line {
-    private String name;
-
-    public Line(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    // 추가 기능 구현
-}
